Add files via upload
Schematic and gerbers for a Gotek(?) compatible with an Artery AT32F435RGT7 chip and a 26 Pin FFC connector.
Used SFRKC30.AT4.35 PCB for most of the schematic and layout but used SFRC2D.B PCB for the 26 Pin FFC connections.
Used photographs and a multimeter along with removing U1 and U3 on the SFRKC30.AT4.35 to get to the traces underneath.
Where possible I followed the original PCB routing of the SFRKC30.AT4.35 but Kicad has some different rules for trace widths (between pins 5&6 of U8 which isn't fitted).
Removed L3 Green LED and changed R21 to 12K to reduce it's brightness. Also increased the pad sizes for the diode D1D to make it easier to hand solder. Minor changes to silk screen.